feat(client): cell view in client (#323)
* feat(client): simulation layer viewer feels solid

* feat(client): big changes to cell view

got rid of it being a dialog since its very awkward blocking view of everything just to see cell contents + couldnt progress the rounds while in it. now appears in game tab. changed a lot of styles of other stuff too

* fix(client): made match statistics collapsable since game tab more crowded

* fix(client): formatting spacing

* chore: need updated spawning mechanics to test w (#319)

* ci(client): update client ci (#314)

* ci: update client ci

* fix: single quotes

* fix: delete first release id

* ci: remove needs lint from release-please

* chore: update echo fro check build

* chore: update win to win32

* chore: remove workflow dispatch

* refactor: update find command

* ci: NOW it only zips the executable

* ci: update release path to work on all platforms

* ci: test the zipping step (will revert)

* Revert "ci: test the zipping step (will revert)"

This reverts commit f6c3189818a6c24e296bd4cdb5ca8638e3873690.

* fix(aegis): make spawns work for dynamic/no dynamic (#310)

* fix(aegis): in theory spawns fixed, client bugged for types=false tho

idk why client still bugged, i think its clients fault

* fix(aegis): fixed spawn zones for agent_types=false

old worlds don't set world.init_spawns, so don't rely on that for a list of all spawn zones

* fix(aegis): game raises error if short agents for spawning

if "positive spawns" are in a world, and there wasn't enough agents to fill out all of these, raise an error, since there is really no reason to run a world like that, and was probably a mistake

* fix(aegis): forgot to not warn if no prio spawns

* fix(client): force blur for input fields when applying brush

* fix(client): force blur for aegis fields too

* fix(client): client console shows errors before games start too

added to see specific error message of "not enough agents", but should also show errors for other "before game" errors too that were previously hidden

* fix(aegis): log more errors on startup

---------

Co-authored-by: Dante <85536418+daannte@users.noreply.github.com>

* fix(client): more styling changes

* refactor(client): TONS of refactoring in client i kinda got carried away

* Fix console buffering (#320)

* fix(all): properly buffered console output in client

* fix(aegis): split agent logging into stdout and stderr again

but it acc seems to be properly flushed now and not out of sync in the client, i think...

* refactor(aegis): forgot unused import mb ruff dont kill me

* fix(client): updated console spacing AGAIN

* refactor(client): agents view refactored to one component

* fix(client): cell view properly updates info as rounds change (i hope)

* style(client): remove grey background

* style(client): little style changes

---------

Co-authored-by: Dante <85536418+daannte@users.noreply.github.com>